Split code into `completion` and `resolve` packages in `toml` module
