+++ toml/src/main/kotlin/org/rust/toml/completion/CargoTomlKnownValuesCompletionProvider.kt
@@ -3,13 +3,15 @@
  * found in the LICENSE file.
  */
 
-package org.rust.toml
+package org.rust.toml.completion
 
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.util.ProcessingContext
+import org.rust.toml.StringValueInsertionHandler
+import org.rust.toml.getClosestKeyValueAncestor
 
 class CargoTomlKnownValuesCompletionProvider(private val knownValues: List<String>) : CompletionProvider<CompletionParameters>() {
     override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
