+++ toml/src/main/kotlin/org/rust/toml/completion/RsTomlKeysCompletionProvider.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.toml
+package org.rust.toml.completion
 
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
@@ -12,6 +12,7 @@ import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.util.ProcessingContext
 import org.intellij.lang.annotations.Language
 import org.rust.lang.core.psi.ext.ancestorStrict
+import org.rust.toml.isDependencyListHeader
 import org.toml.lang.psi.*
