+++ toml/src/test/kotlin/org/rust/toml/completion/CargoTomlCompletionEditionTest.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.toml
+package org.rust.toml.completion
 
 import org.junit.After
 import org.junit.Before
