+++ toml/src/test/kotlin/org/rust/toml/resolve/CargoTomlResolveTestBase.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.toml
+package org.rust.toml.resolve
 
 import org.rust.FileTreeBuilder
 import org.rust.fileTree
